Spotify: keep the artist type out of the component layer
