+++ frontend/components/prospect-profile-screen/prospect-profile-screen.tsx
@@ -76,11 +76,8 @@ import Reanimated, {
   withTiming,
 } from 'react-native-reanimated';
 import { ClubItem, joinClub, leaveClub } from '../../club/club';
-import {
-  SpotifyArtistItem,
-  SpotifyArtists,
-  SpotifyTitle,
-} from '../spotify-artists';
+import { SpotifyArtistItem } from '../../api/spotify';
+import { SpotifyArtists, SpotifyTitle } from '../spotify-artists';
 import * as _ from 'lodash';
 import {
   friendlyTimeAgo,
